+++ robocorp-code/codegen/views.py
@@ -205,6 +205,11 @@ TREE_VIEW_CONTAINERS = [
                             MenuGroup.INLINE,
                             "robocorp-code:single-robot-selected && viewItem == newWindowsLocator",
                         ),
+Menu(
+                            "robocorp.newRobocorpInspectorRecord",
+                            MenuGroup.INLINE,
+                            "robocorp-code:single-robot-selected && viewItem == newRecord",
+                        ),
                         # Locators (root)
                         Menu(
                             "robocorp.openLocatorsJson",